requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2
matplotlib==3.8.4
streamlit==1.35.0
plotly==5.22.0
fpdf2==2.7.9
python-multipart==0.0.9
aiofiles==23.2.1
